feat: better exception handling
move AJAX handling to the global exception handler
fixing/workaround some problems on client site when replace the whole DOM

issue: TOBAGO-1946
